requirements.txt
# Core dependencies
langgraph==0.2.56
langchain==0.3.14
langchain-openai==0.2.13
langchain-anthropic==0.3.0
pydantic==2.10.3
pydantic-settings==2.6.1

# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==13.1
httpx==0.28.1

# Database
aiosqlite==0.20.0
sqlalchemy==2.0.36

# MCP Client
mcp==1.3.8

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
tenacity==9.0.0
structlog==24.4.0
rich==13.9.4

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.15.0
black==24.10.0
ruff==0.8.4
mypy==1.14.0
pre-commit==4.0.1

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.49